fix: recursive trait bound for `Chrono04Clock` and `StdClock`
During the renames performed for version `0.2.0`, the trait `StdClock` was renamed to `StdClock`, and the trait `StdScheduler` was renamed to `StdClock`. However, trait bounds were missed during the first rename, this caused the circular bound `StdClock: StdClock`. A similar issue happened with the `Chrono04*` traits. This commit fixes this broken rename.